requirements.txt
beautifulsoup4==4.8.2
boto==2.49.0
celery==4.4.0
Django==2.2.9
django-cors-headers==3.2.1
django-extensions==2.2.5
django-mysql==3.3.0
django-organizations==1.1.2
django-storages==1.8
djangorestframework==3.11.0
django-simple-history==2.8.0
django-redis==4.11.0
drf-api-tracking==1.6.0
filetype==1.0.5
ipython
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.9.1
simplejson==3.17.0
unicodecsv==0.14.1
xlrd==1.2.0
constants==0.6.0
django-jsonfield==1.4.0
django-user-accounts==2.1.0
django-import-export==2.0.2
django-unused-media==0.1.13
python-magic==0.4.15
pysftp==0.2.9
pinax-theme-bootstrap==8.0.1
pandas==0.25.3
editdistance==0.5.3
openpyxl==3.0.3
Unidecode==1.0.23
